LayoutTests/imported/w3c/web-platform-tests/css/css-inline/baseline-source/baseline-source-first-textarea-003.tentative-expected.txt








FAIL .target > * 1 assert_equals:
<span data-offset-x="30"></span>
offsetLeft expected 30 but got 45
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<span data-offset-x="30"></span>
offsetLeft expected 30 but got 60
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<span data-offset-x="30"></span>
offsetLeft expected 30 but got 60
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<span data-offset-x="65"></span>
offsetLeft expected 65 but got 60
PASS .target > * 8
FAIL .target > * 9 assert_equals:
<span data-offset-x="65"></span>
offsetLeft expected 65 but got 45
PASS .target > * 10
FAIL .target > * 11 assert_equals:
<span data-offset-x="65"></span>
offsetLeft expected 65 but got 45
PASS .target > * 12
